requirements.txt
# --- Core backend ---
fastapi==0.111.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.0
python-multipart==0.0.9
pymongo[srv]==4.7.2

# --- Auth & security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- AI & utilities ---
google-generativeai>=0.8.0
httpx==0.25.2

# --- File processing ---
Pillow==10.4.0
python-docx==1.1.0

# --- Cloud storage ---
cloudinary==1.41.0

# --- (Optional) PDF generation ---
reportlab==4.0.7

# test
# --- Dev only (KHÔNG dùng khi deploy) ---
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-mock==3.11.1
pydantic-settings==2.4.0
pydantic==2.9.2
sqlmodel==0.0.16
